/archives returns all entries rather than 0
On /archives, 0 entries were returned because the filter is not
populated with a date string, so no entries matched.  The filter in this
case is populated with the request type.  To return all entries on
/archives, check if the filter matches a date string.  If the filter
matches a date string, select entries with the filter, otherwise select
all entries.

This patch considers the filter to match a date string if its first four
characters are digits.