Fix 'time out of range' bug for ancient / future dates.
On my 32bit linux system with ruby 1.8.7 I get a 'time out of range'
exceptionif the date is before 1901-12-14 or after 2038-01-18.
Turns out the Date class has lower resolution, and therefore wider
available date range.

This would seem like an unlikely bug, but since the year of the
example article in the dorothy template as of writing is set to 1900,
it immediately struck.