Use date from filename to generate url.
Keep things DRY: Override the date found in the metadata with the
date from the filename. If a date is given in the metadata, that
date will show up on the article page, but the url will use the date
from the filename.

This is needed to make sure toto actually finds the articles, which
it would previously not if the date in the metadata did not match
the date from the filename. In fact, with this patch it is no longer
necessary to put the date in the article metadata at all, you only
need it in the filename, which keep things DRY.